backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.*
python-dotenv==1.*
sqlalchemy>=2.0
psycopg2-binary
requests
httpx
readability-lxml
lxml
google-generativeai